+++ packages/analytics-core/test/timeline.test.ts
@@ -1,9 +1,16 @@
-import { register, deregister, push, apply, flush } from '../src/timeline';
-import { AmplitudeDestinationPlugin, Event, Plugin, PluginType } from '@amplitude/analytics-types';
+import { Timeline } from '../src/timeline';
+import { Event, Plugin, PluginType } from '@amplitude/analytics-types';
 import { OPT_OUT_MESSAGE } from '../src/messages';
 import { useDefaultConfig } from './helpers/default';
+import { createTrackEvent } from '../src/utils/event-builder';
 
 describe('timeline', () => {
+  let timeline = new Timeline();
+
+  beforeEach(() => {
+    timeline = new Timeline();
+  });
+
   test('should update event using before/enrichment plugin', async () => {
     const beforeSetup = jest.fn().mockReturnValue(Promise.resolve());
     const beforeExecute = jest.fn().mockImplementation((event: Event) =>
@@ -55,35 +62,41 @@ describe('timeline', () => {
     };
     const config = useDefaultConfig();
     // register
-    await register(before, config);
-    await register(enrichment, config);
-    await register(destination, config);
+    await timeline.register(before, config);
+    await timeline.register(enrichment, config);
+    await timeline.register(destination, config);
     expect(beforeSetup).toHaveBeenCalledTimes(1);
     expect(enrichmentSetup).toHaveBeenCalledTimes(1);
     expect(destinationSetup).toHaveBeenCalledTimes(1);
-    expect(config.plugins.length).toBe(3);
+    expect(timeline.plugins.length).toBe(3);
     const event = (id: number): Event => ({
       event_type: `${id}:event_type`,
     });
     await Promise.all([
-      push(event(1), config).then(() => push(event(1.1), config)),
-      push(event(2), config).then(() => Promise.all([push(event(2.1), config), push(event(2.2), config)])),
-      push(event(3), config).then(() =>
-        Promise.all([
-          push(event(3.1), config).then(() => Promise.all([push(event(3.11), config), push(event(3.12), config)])),
-          push(event(3.2), config),
-        ]),
-      ),
+      timeline.push(event(1), config).then(() => timeline.push(event(1.1), config)),
+      timeline
+        .push(event(2), config)
+        .then(() => Promise.all([timeline.push(event(2.1), config), timeline.push(event(2.2), config)])),
+      timeline
+        .push(event(3), config)
+        .then(() =>
+          Promise.all([
+            timeline
+              .push(event(3.1), config)
+              .then(() => Promise.all([timeline.push(event(3.11), config), timeline.push(event(3.12), config)])),
+            timeline.push(event(3.2), config),
+          ]),
+        ),
     ]);
     expect(beforeExecute).toHaveBeenCalledTimes(10);
     expect(enrichmentExecute).toHaveBeenCalledTimes(10);
     expect(destinationExecute).toHaveBeenCalledTimes(10);
 
     // deregister
-    await deregister(before.name, config);
-    await deregister(enrichment.name, config);
-    await deregister(destination.name, config);
-    expect(config.plugins.length).toBe(0);
+    await timeline.deregister(before.name);
+    await timeline.deregister(enrichment.name);
+    await timeline.deregister(destination.name);
+    expect(timeline.plugins.length).toBe(0);
   });
 
   describe('push', () => {
@@ -93,7 +106,7 @@ describe('timeline', () => {
       };
       const config = useDefaultConfig();
       config.optOut = true;
-      const results = await push(event, config);
+      const results = await timeline.push(event, config);
       expect(results).toEqual({
         event,
         code: 0,
@@ -103,7 +116,7 @@ describe('timeline', () => {
   });
 
   describe('apply', () => {
-    test('should handle empty queue', async () => {
+    test('should handle undefined event', async () => {
       const beforeSetup = jest.fn().mockReturnValueOnce(Promise.resolve());
       const beforeExecute = jest.fn().mockImplementationOnce((event: Event) =>
         Promise.resolve({
@@ -118,69 +131,36 @@ describe('timeline', () => {
         execute: beforeExecute,
       };
       const config = useDefaultConfig();
-      await register(before, config);
-      await apply();
-      await deregister(before.name, config);
+      await timeline.register(before, config);
+      await timeline.apply(undefined);
+      await timeline.deregister(before.name);
       expect(beforeExecute).toHaveBeenCalledTimes(0);
     });
   });
 
   describe('flush', () => {
-    test('should call destination flush method', async () => {
-      const destinationSetup = jest.fn().mockReturnValue(Promise.resolve());
-      const destinationExecute = jest
-        .fn()
-        // error once
-        .mockImplementationOnce((event: Event) => {
-          expect(event.event_id).toBe('1');
-          expect(event.user_id).toBe('2');
-          return Promise.reject({});
-        })
-        // success for the rest
-        .mockImplementation((event: Event) => {
-          expect(event.event_id).toBe('1');
-          expect(event.user_id).toBe('2');
-          return Promise.resolve();
-        });
-      const destinationFlush = jest.fn().mockReturnValue(Promise.resolve());
-      const destination: AmplitudeDestinationPlugin = {
-        name: 'plugin:destination',
+    test('should flush events', async () => {
+      const setup = jest.fn().mockReturnValueOnce(Promise.resolve(undefined));
+      const execute = jest.fn().mockReturnValue(Promise.resolve(undefined));
+      const flush = jest.fn().mockReturnValue(Promise.resolve(undefined));
+      const plugin = {
+        name: 'mock',
         type: PluginType.DESTINATION,
-        setup: destinationSetup,
-        execute: destinationExecute,
-        flush: destinationFlush,
+        setup,
+        execute,
+        flush,
       };
       const config = useDefaultConfig();
-      await register(destination, config);
-      await flush(config);
-      expect(destinationFlush).toHaveBeenCalledTimes(1);
-    });
-
-    test('should not throw error if destination plugin does not have flush method', async () => {
-      const destinationSetup = jest.fn().mockReturnValue(Promise.resolve());
-      const destinationExecute = jest
-        .fn()
-        // error once
-        .mockImplementationOnce((event: Event) => {
-          expect(event.event_id).toBe('1');
-          expect(event.user_id).toBe('2');
-          return Promise.reject({});
-        })
-        // success for the rest
-        .mockImplementation((event: Event) => {
-          expect(event.event_id).toBe('1');
-          expect(event.user_id).toBe('2');
-          return Promise.resolve();
-        });
-      const destination: Plugin = {
-        name: 'plugin:destination',
-        type: PluginType.DESTINATION,
-        setup: destinationSetup,
-        execute: destinationExecute,
-      };
-      const config = useDefaultConfig();
-      await register(destination, config);
-      await expect(flush(config)).resolves.not.toThrowError();
+      await timeline.register(plugin, config);
+      void timeline.push(createTrackEvent('a'), config);
+      void timeline.push(createTrackEvent('b'), config);
+      void timeline.push(createTrackEvent('c'), config);
+      void timeline.push(createTrackEvent('d'), config);
+      expect(timeline.queue.length).toBe(4);
+      await timeline.flush();
+      expect(timeline.queue.length).toBe(0);
+      expect(execute).toBeCalledTimes(4);
+      expect(flush).toBeCalledTimes(1);
     });
   });
 });