+++ packages/analytics-types/src/destination-context.ts
@@ -5,5 +5,5 @@ export interface DestinationContext {
   event: Event;
   attempts: number;
   callback: EventCallback;
-  delay: number;
+  timeout: number;
 }